requirements.txt
azure-common==1.1.28
azure-storage-blob==12.16.0
bleach==6.0.0
boto3==1.26.152
celery==5.3.0
Django==4.2.2
django-celery-beat==2.5.0
django-solo==2.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
numpy==1.24.3
openpyxl==3.1.2
pandas==2.0.2
Pillow==9.5.0
psycopg2==2.9.6
Pygments==2.15.1
python-magic==0.4.27
redis==4.5.5
requests==2.31.0
sentry-sdk==1.25.1
tqdm==4.65.0
ua-parser==0.16.1
